Uno R4 incompatibility
Uno R4 Wifi/Minima had an incompatibility which is fixed here. Incremented version and prepared for new release.

+++ src/PTSolns_InterfaceShield.h
@@ -144,7 +144,7 @@ class Interface : public Print
 
     unsigned long btnTimes[4] = {0, 0, 0, 0};
     bool btnStates[4] = {LOW, LOW, LOW, LOW};
-    int debounceTime = 100;
+    unsigned long debounceTime = 100;
     bool debounceEnabled = false;
 
     union Ports
